rpm: Give up on attempting to separate controlled and uncontrolled files
SuiteCRM makes extremely liberal use of __FILE__ and __DIR__ in
several hundred places.  The morons responsible for the PHP language
chose to have these resolve symlinks automatically, thereby providing
PHP code with literally no sensible way to determine the path through
which it was originally accessed.

The consequences of this particular idiocy are present everywhere
within SuiteCRM.  Data (including private keys) is stored in the same
directories as source code.  Self-modifying autogenerated code is
stored in the same directories as source code.  The sheer level of
stupidity involved in creating something like this is astounding.

Give up on the idea of having a single set of SuiteCRM code files in
/usr/share and symlinks in /var/www (as is done for e.g. MediaWiki),
since the SuiteCRM code is far too poorly written to cope with this.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>