rpm: Allow for multi-instance systemd services
Restructure to allow for multiple systemd instances, e.g.

  systemctl start suitecrm@sales
  systemctl start suitecrm@logistics

This approach uses systemd's StateDirectory, CacheDirectory, etc
directives to define the per-instance directories, and relies on an
ExecStartPre script to create the required subdirectories and symlinks
for each instance.  This removes a reasonable amount of complexity
from the specfile, which now simply claims ownership of the top-level
/var/{lib,log,cache}/suitecrm and /etc/suitecrm directories.

Since systemd is now aware of the various directories, it becomes
possible to run commands such as

  systemctl clean suitecrm@sales

to clear out SuiteCRM's cache directory (which is often required to
work around bugs in SuiteCRM).

A dedicated php-fpm process is used for each instance, since there
seems to be no way to set the PHPTURD environment variable on a
per-request basis.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>