rpm: Block access to .log files within cache directory
The SuiteCRM autogenerated .htaccess file (which we do not use) blocks
access to all .log files.  Our package already separates out the logs
to /var/log/suitecrm, which is not accessible via the web server.

It is unclear whether or not .log files are ever written to the cache
directory by random parts of SuiteCRM code (or external modules).  Err
on the side of caution by blocking them.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>