rpm: Work around an apparent systemd bug in detecting policy reloads
Restore the selinux file contexts on the directories created by
systemd as part of service initialisation.

This ought not to be necessary.  systemd's selinux-util.c installs a
callback that is supposed to detect an selinux policy reload.
However, experiments with strace show that systemd receives the
SELNL_MSG_POLICYLOAD but then completely ignores the updated policy.
Using "systemctl daemon-reexec" causes systemd to pick up the new
policy, but this is not something we want to do from an RPM
postinstall script.

Work around the problem by using restorecon to correct the file
contexts on the directories that were created by systemd based on
potentially-outdated policy.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>