rpm: Make config_override.php a writable file
Contrary to the documentation, the config_override.php file is the one
that SuiteCRM will update to reflect configuration changes made
through the web interface.  If the file is not writable, then
configuration changes are silently discarded.

Make config_override.php a writable file to minimise surprise.  Leave
config.php as read-only since the SuiteCRM installer will handle that
gracefully by falling back to providing the contents for the
administrator to copy and paste.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>