rpm: Allow for inclusion of additional configuration fragments
Allow for configuration fragments placed in a conf.d directory to be
assembled into a conf_d.php that is included after config.php and
config_override.php.

Remove the ExecReload since the reload mechanism would not pick up
changes to these configuration file fragments, which would be
unpleasantly surprising for a user.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>