rpm: Work around a segfault in systemd-sysusers 239 and earlier
There is a bug in systemd version 239 and earlier which causes a
segfault if systemd-sysusers is invoked with the --replace option
specifying a file that actually exists.  This bug was fixed in systemd
commit

  https://github.com/systemd/systemd/commit/082bb1c

Work around the bug on CentOS (which ships with systemd 239) by
deleting the file before invoking systemd-sysusers.  This is harmless
since:

- deleting the file achieves the same effect as --replace in terms of
  the expected functionality from running systemd-sysusers

- the file will be immediately replaced in the subsequent stage of
  package installation anyway

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>