rpm: Document use of httpd_can_network_connect
Some functionality (such as sending emails) requires the ability to
connect to external servers.  This is blocked by the default selinux
security policy.

It is debatable whether or not this default policy is appropriate, but
the installation of a package should not silently relax a security
policy that was presumably a deliberate decision.

The best solution seems to be to document the likely requirement to
use "setsebool -P httpd_can_network_connect on".  This is already a
common pattern familiar to most system administrators.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>